% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_data_frame.R
\name{from_data_frame}
\alias{from_data_frame}
\title{Convert a dataframe to a bedtools object}
\usage{
from_data_frame(df)
}
\arguments{
\item{df}{A bedtools object that points to a file, and should have been generated by the bedtools command}
}
\value{
A bedtools object
}
\description{
Convert a dataframe to a bedtools object
}
