\name{rBedtools-package}
\alias{rBedtools-package}
\alias{rBedtools}
\docType{package}
\title{
\packageTitle{rBedtools}
}
\description{
\packageDescription{rBedtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rBedtools}
\packageIndices{rBedtools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{rBedtools}

Maintainer: \packageMaintainer{rBedtools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
