% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBedtools.R
\name{RBedtools}
\alias{RBedtools}
\title{Main interface to bedtools}
\usage{
RBedtools(tool, options = "", output = tmp_loc(), ...)
}
\arguments{
\item{tool}{which bedtools function to use}

\item{options}{A character vector containing options to use for tools. use bedtools intersec -h on CL to see options}

\item{output}{can be either a path to a file to write to, or 'stdout' to direct output to new bedtools command}

\item{...}{Arguments to pass to bed tools that are specfic to tool}
}
\value{
a bedtools object that points to an output file, or bedtools command object to add to another command
}
\description{
Main interface to bedtools
}
