% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_data_frame.R
\name{to_data_frame}
\alias{to_data_frame}
\title{Convert bedtools object to dataframe}
\usage{
to_data_frame(obj)
}
\arguments{
\item{obj}{A bedtools object that points to a file, and should have been generated by the bedtools command}
}
\value{
A data.frame
}
\description{
Convert bedtools object to dataframe
}
